feat(PackageImporter): prompt before adding scoped registry
It is too heavy handed to automatically add a scoped registry and
change the Unity project manifest.json file without the user's approval
so instead of automatically adding it, if the scoped registry is not
found then now a warning box appears with a button to click that
performs the action of adding the scoped registry. This is then at
the user's choice.