feat(PackageImporter): add multi import checkbox next to package names
It is now possible to import multiple packages at the same time by
ticking the checkbox next to the package name and clicking the
`Add Selected Packages` button at the bottom of the list.

This feature only works in Unity 2021.2 and above due to the inclusion
of the `Client.AddAndRemove` method that is not available in previous
versions of the Unity software.

And due to how Unity will reload after a package is added, it means
there is no easy way of keeping track of a list of packages without
resorting to the editor prefs to store the list, which feels like
overkill.